requirements.txt
fastapi
uvicorn[standard]
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12,<5.0
PyJWT>=2.9,<3.0
pandas>=2.0
streamlit>=1.37
altair==5.3.0
typing_extensions>=4.12
